Release point for 0.5 version
Compilation warning fix

git-svn-id: https://glfx.googlecode.com/svn/trunk@15 5595c9b7-382c-e053-034d-f28d15879402

+++ src/glfxScanner.cpp
@@ -1550,7 +1550,7 @@ if(gLexPassthrough) gEffect->m_sharedCode<<glfxtext; else {glfxlval.num=atoi(glf
 case 13:
 YY_RULE_SETUP
 #line 75 "glfx.lpp"
-if(gLexPassthrough) gEffect->m_sharedCode<<glfxtext; else {glfxlval.fnum=atof(glfxtext); return FLOAT;}
+if(gLexPassthrough) gEffect->m_sharedCode<<glfxtext; else {glfxlval.fnum=(float)atof(glfxtext); return FLOAT;}
 	YY_BREAK
 case 14:
 YY_RULE_SETUP
